+++ energyplus/src/main/java/com/kh/ecolog/market/model/service/MarketServiceImpl.java
@@ -5,9 +5,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.kh.ecolog.auth.model.vo.CustomUserDetails;
+import com.kh.ecolog.auth.service.AuthService;
+import com.kh.ecolog.common.util.SecurityUtil;
 import com.kh.ecolog.file.service.FileService;
 import com.kh.ecolog.market.model.dao.MarketMapper;
 import com.kh.ecolog.market.model.dto.MarketDTO;
@@ -22,10 +27,9 @@ import lombok.extern.slf4j.Slf4j;
 public class MarketServiceImpl implements MarketService  {
 	private final MarketMapper marketMapper;
 	private final FileService fileService;
-	
-	
 
 	private void saveMarketImages(Long marketNo, List<MultipartFile> images) {
+		
 	    if (images == null || images.size() != 3) {
 	        throw new IllegalArgumentException("이미지는 3장 등록해야 합니다.");
 	    }
@@ -47,6 +51,13 @@ public class MarketServiceImpl implements MarketService  {
 	
 	@Override
 	public void insertMarket(MarketDTO dto, List<MultipartFile> images) {
+		
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+	    CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
+	    
+	    Long userId = user.getUserId();
+	    dto.setUserId(userId);
+	    
 	    dto.setMarketStatus("N");
 	    dto.setMarketDate(new Date(System.currentTimeMillis()));
 
@@ -56,6 +67,9 @@ public class MarketServiceImpl implements MarketService  {
 
 	@Override
 	public void updateMarket(MarketDTO dto, List<MultipartFile> images) {
+		
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+	    CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
 
 	    // 1. 유효한 새 이미지만 필터링
 	    List<MultipartFile> validImages = (images != null) 
@@ -121,7 +135,17 @@ public class MarketServiceImpl implements MarketService  {
 	
 	@Override
 	public void deleteMarket(Long marketNo, Long userId) {
-		 // 1. 관련 이미지 먼저 삭제
+		
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+	    CustomUserDetails user = (CustomUserDetails) auth.getPrincipal();
+		
+		// 1. 글 작성자 확인
+	    MarketDTO market = marketMapper.selectMarketByNo(marketNo);
+	    if (!market.getUserId().equals(userId)) {
+	        throw new SecurityException("작성자만 삭제할 수 있습니다.");
+	    }
+	    
+	    
 		marketMapper.deleteImagesByMarketNo(marketNo);
 
 	    // 2. 게시글 삭제
